refactor(voice-call): share proxy IP normalization
